Move Transformation to its own file, transform.py (#218)
