+++ galsim/__init__.py
@@ -130,8 +130,8 @@ from interpolatedimage import Interpolant
 from interpolatedimage import Nearest, Linear, Cubic, Quintic, Lanczos, SincInterpolant, Delta
 from interpolatedimage import InterpolatedImage
 from compound import Add, Sum, Convolve, Convolution, Deconvolve, Deconvolution
-from compound import Transform, Transformation
 from compound import AutoConvolve, AutoConvolution, AutoCorrelate, AutoCorrelation
+from transform import Transform, Transformation
 
 # Chromatic
 from chromatic import ChromaticObject, ChromaticAtmosphere, Chromatic, ChromaticSum
